refactor: modularize response writers
